+++ elements/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- [pos-list](../docs/elements/components/pos-list):
+  - Basic implementation of element to list things related to a resource with a custom display
+
 ## 0.25.2
 
 ### Fixed